+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [Unreleased]
+
+* Add form block importer script
+
 ### 1.0.6: 2022-02-22
 
 * Fix adding block class to new block